fixed a typo that caused zone to not compile.

+++ zone/special_attacks.cpp
@@ -641,7 +641,7 @@ void Mob::RogueBackstab(Mob* other, bool min_damage, int ReuseTime)
 		ndamage = -5;
 	}
 	
-	if (IsClient() && CastToClient()->GetBuildRank(ROGUE, RB_ROG_VITALORGANS) > 0 && zone->random.Roll(10 * CastToClient()->GetBuildRank(ROGUE, RB_ROG_VITALORGANS))) {
+	if (IsClient() && CastToClient()->GetBuildRank(ROGUE, RB_ROG_VITALORGANS) > 0 && zone->random.Roll((int)(10 * CastToClient()->GetBuildRank(ROGUE, RB_ROG_VITALORGANS)))) {
 		ndamage += ndamage * 0.25f * CastToClient()->GetBuildRank(ROGUE, RB_ROG_VITALORGANS);
 	}
